refactor: simplify locale routing
