--- src/app/[locale]/decks/[slug]/opengraph-image.tsx
@@ -1,33 +1,0 @@
-import { ImageResponse } from "next/og";
-
-import { getDeckBySlug } from "@/lib/deck-store";
-import { renderDeckSocialImage } from "@/lib/social-image";
-
-export const runtime = "nodejs";
-export const size = {
-  width: 1200,
-  height: 630,
-};
-export const contentType = "image/png";
-
-export default async function DeckOpenGraphImage({
-  params,
-}: {
-  params: Promise<{ locale: string; slug: string }>;
-}) {
-  const { locale, slug } = await params;
-  const record = await getDeckBySlug(slug);
-  const detailPath = `/${locale}/decks/${slug}`;
-
-  return new ImageResponse(
-    renderDeckSocialImage(record?.metadata, {
-      locale,
-      width: size.width,
-      height: size.height,
-      detailPath,
-    }),
-    {
-      ...size,
-    },
-  );
-}
