--- src/app/[locale]/layout.tsx
@@ -1,44 +1,0 @@
-export const runtime = "nodejs";
-
-import { NextIntlClientProvider } from "next-intl";
-import { notFound } from "next/navigation";
-import { setRequestLocale } from "next-intl/server";
-import type { ReactNode } from "react";
-
-import { SiteFooter } from "@/components/site-footer";
-import { SiteHeader } from "@/components/site-header";
-import { locales, type AppLocale } from "@/i18n/config";
-import { getMessages } from "@/i18n/get-messages";
-
-export function generateStaticParams() {
-  return locales.map((locale) => ({ locale }));
-}
-
-export default async function LocaleLayout({
-  children,
-  params,
-}: {
-  children: ReactNode;
-  params: Promise<{ locale: string }>;
-}) {
-  const { locale } = await params;
-
-  if (!locales.includes(locale as AppLocale)) {
-    notFound();
-  }
-
-  const appLocale = locale as AppLocale;
-  setRequestLocale(appLocale);
-
-  const messages = await getMessages(appLocale);
-
-  return (
-    <NextIntlClientProvider locale={appLocale} messages={messages} timeZone="UTC">
-      <div className="flex min-h-screen flex-col">
-        <SiteHeader />
-        <main className="flex-1">{children}</main>
-        <SiteFooter />
-      </div>
-    </NextIntlClientProvider>
-  );
-}
