--- src/app/[locale]/twitter-image.tsx
@@ -1,31 +1,0 @@
-import { ImageResponse } from "next/og";
-
-import { renderDeckSocialImage } from "@/lib/social-image";
-
-export const runtime = "nodejs";
-export const size = {
-  width: 800,
-  height: 418,
-};
-export const contentType = "image/png";
-
-export default async function LocaleTwitterImage({
-  params,
-}: {
-  params: Promise<{ locale: string }>;
-}) {
-  const { locale } = await params;
-  const detailPath = `/${locale}`;
-
-  return new ImageResponse(
-    renderDeckSocialImage(undefined, {
-      locale,
-      width: size.width,
-      height: size.height,
-      detailPath,
-    }),
-    {
-      ...size,
-    },
-  );
-}
