+++ src/i18n/config.ts
@@ -3,7 +3,7 @@ export type AppLocale = (typeof locales)[number];
 
 export const defaultLocale: AppLocale = "en";
 
-export const localePrefix = "always" as const;
+export const localePrefix = "never" as const;
 
 export const LOCALE_LABELS: Record<AppLocale, string> = {
   en: "English",
